v0.972: RP recovery recommendations!
tailoring recommendations to ITL point value: in my mind, the player should be able to "access" whatever EX score moves the point to the trend line, and from that, whatever additional ITL points are generated. then I can rank charts based on that increment, as well as whether or not they would be in top 75 if they were raised to that score.

recs are reported as
- current %EX score
- target %EX score
- gain in ITL points for that chart
- increase in ITL RP resulting from raising that chart in isolation

also in this update:
- plots now include recency (saturation of pink) and replay (size of dot) information